Lots of changes and additions have been made.  Weight class was added, player and weight classes received a collision mask.  Collisions between player and weight are handled and player is now able to throw the weight and recatch it without issue.  The weight can fall and hit the floor without falling through it.  Gravity was added to the weight class.  Almost every class in some way has been modified to some degree in this latest update.

+++ core/src/helpers/GameInfo.java
@@ -5,6 +5,6 @@ package helpers;
  */
 
 public class GameInfo {
-    public static int WIDTH = 720;
-    public static int HEIGHT = 1280;
+    public static float WIDTH = 720;
+    public static float HEIGHT = 1280;
 }
